requirements.txt
flask
uvicorn 
python-dotenv
sqlalchemy
psycopg2
fastapi